+++ lib/wavefront-cli/commands/window.rb
@@ -16,7 +16,7 @@ class WavefrontCommandWindow < WavefrontCommandBase
   end
 
   def _commands
-    ["list #{CMN} [-al] [-f format] [-o offset] [-L limit]",
+    ["list #{CMN} [-al] [-O fields] [-f format] [-o offset] [-L limit]",
      "describe #{CMN} [-f format] <id>",
      "create #{CMN} -d reason [-s time] [-e time] " \
      '[-A alert_tag...] [-T host_tag...] [-H host...] <title>',
@@ -32,16 +32,17 @@ class WavefrontCommandWindow < WavefrontCommandBase
 
   def _options
     [common_options,
-     '-l, --long           list maintenance windows in detail',
-     '-a, --all            list all maintenance windows',
-     '-o, --offset=n       start from nth maintenance window',
-     '-L, --limit=COUNT    number of maintenance windows to list',
-     '-d, --desc=STRING    reason for maintenance window',
-     '-s, --start=TIME     time at which window begins',
-     '-e, --end=TIME       time at which window ends',
-     '-A, --atag=STRING    alert tag to which window applies',
-     '-H, --host=STRING    host to which window applies',
-     '-T, --htag=STRING    host tag to which window applies',
-     '-f, --format=STRING  output format']
+     '-l, --long              list maintenance windows in detail',
+     '-a, --all               list all maintenance windows',
+     '-o, --offset=n          start from nth maintenance window',
+     '-O, --fields=F1,F2,...  only show given fields',
+     '-L, --limit=COUNT       number of maintenance windows to list',
+     '-d, --desc=STRING       reason for maintenance window',
+     '-s, --start=TIME        time at which window begins',
+     '-e, --end=TIME          time at which window ends',
+     '-A, --atag=STRING       alert tag to which window applies',
+     '-H, --host=STRING       host to which window applies',
+     '-T, --htag=STRING       host tag to which window applies',
+     '-f, --format=STRING     output format']
   end
 end