+++ spec/wavefront-cli/commands/message_spec.rb
@@ -11,6 +11,6 @@ require_relative 'base_spec'
 class WavefrontCommmandMessageTest < WavefrontCommmandBaseTest
   def setup
     @wf = WavefrontCommandMessage.new
-    @col_width = 19
+    @col_width = 22
   end
 end